+++ lib/trace_location.rb
@@ -15,8 +15,8 @@ module TraceLocation # :nodoc:
     Report.build(result.events, result.return_value, options).generate
     true
   rescue StandardError => e
-    $stderr.puts "Failure: TraceLocation got an unexpected error."
-    $stderr.puts e.full_message
+    warn 'Failure: TraceLocation got an unexpected error.'
+    warn e.full_message
     false
   end